Escape needed characters to calculate expectedHash
Before calculating the hash that is sent by facebook it escapes some special characters,
so to verify if the incoming hash is right it is needed to also escape the same characters.

This code maps some special characters that needed escaping.

More about this on [facebook](https://developers.facebook.com/docs/messenger-platform/webhook/#security)

Possible enhancements of the code are:
* use a Map instead of an ordinary object to map the characters
* check efficiency of executing multiple replaces at the same time using [regex](https://stackoverflow.com/questions/15604140/replace-multiple-strings-with-multiple-other-strings)